+++ src/fastify/index.ts
@@ -1,2 +1,4 @@
+export * from './cache';
 export * from './fastify';
 export * from './openapi';
+export * from './schemas';